requirements.txt
hypercorn==0.17.3
python-dotenv==1.0.1
quart==0.19.9
quart_cors==0.7.0
Requests==2.32.3
Telethon==1.38.1
tinydb==4.8.2
